Update DynamicArrayQueue.java
In this DynamicArrayQueue

1) When a user is providing capacity for the queue we are not resetting the CAPACITY  variable in our constructor function's cap value.
    we are using our CAPACITY variable further in the code to increment rear and front, so if the actual capacity is not taken into account then we can get a wrong index.

The existing code for the expand function is not working for a circular queue scenario.
//This loop wont work for circular case when our rear will be pointing to the same location as that of front , i.e 
		// queue[5] = {1,2,3,4,5} , now our rear will be at (rear + 1) % CAPACITY
		// i.e 0 and loop condition will run only once.
// 		for(int i = front; i <= rear; i ++)
// 			newQueue[i-front] = queueRep[i%CAPACITY];


rear = size also has an issue
If we go up till size - 1 then we will be pointing the latest element entered in our queue and not the latest free slot to enter the value rear = size-1. As the size is incremented each time on enqueue and decremented each time at dequeue.